backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==44.0.0
email-validator==2.2.0
